// schemas/model/eventLogResult.ts
/**
 * Generated by orval v6.23.0 🍺
 * Do not edit manually.
 * My API
 * OpenAPI spec version: 1.0.0
 */
import type { Event } from './event';
import type { EventLogResultExtra } from './eventLogResultExtra';

export interface EventLogResult {
  comment?: string;
  createdAt: string;
  event: Event;
  eventName: string;
  eventTime?: string;
  extra?: EventLogResultExtra;
  id: number;
  updatedAt: string;
}